pyproject.toml
[project]
name = "omnimise-scholarship-agent"
version = "0.1.0"
description = "Scholarship scraper, user profiles, and matching API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
  "sqlalchemy>=2.0.36",
  "alembic>=1.14.0",
  "pydantic>=2.10.0",
  "pydantic-settings>=2.6.0",
  "httpx>=0.28.0",
  "beautifulsoup4>=4.12.0",
  "apscheduler>=3.10.4",
  "python-multipart>=0.0.12",
  "pypdf>=5.1.0",
  "python-docx>=1.1.2",
  "python-dateutil>=2.9.0",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.8.0",
]

[build-system]
requires = ["setuptools>=75.0.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
